Chapter 2

+++ Pytorch Book/Chapter 1/perceptron.py
@@ -9,7 +9,7 @@ class Perceptron:
         rgen = np.random.RandomState(self.random_state)
         self.w_ = rgen.normal(loc=0.0, scale=0.01,size=X.shape[1])
 
-        self.b_ = np.float_(0.)
+        self.b_ = float(0.)
         self.errors_ = []
